[pulsar-broker] broker resources use metadata-store api
fix test
